<REFACTOR>(OPMODES): disabled and deprecated unused opmods
<REFACTOR>(AUTO): deprecate rightauto
