+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/experimental/Experimental_Auto_Integration.java
@@ -7,10 +7,10 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.opmodes.RightAuto;
 import org.firstinspires.ftc.teamcode.roadrunner.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.subsystems.Belt;
 import org.firstinspires.ftc.teamcode.subsystems.Camera;
@@ -21,6 +21,8 @@ import org.firstinspires.ftc.teamcode.subsystems.TurnTable;
 
 @Autonomous(name = "Integrated_Auto")
 @Config
+@Disabled
+
 public class Experimental_Auto_Integration extends LinearOpMode {
 
     // facing field side
