+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/experimental/Experimental_StationaryScore.java
@@ -7,6 +7,7 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -20,6 +21,8 @@ import org.firstinspires.ftc.teamcode.subsystems.TurnTable;
 
 @Autonomous(name = "StationaryEx")
 @Config
+@Disabled
+
 public class Experimental_StationaryScore extends LinearOpMode {
 
     // facing field side
